requirements.txt
torch
diffusers
transformers
Pillow
grpcio
grpcio-tools
gradio
